requirements.txt
python-telegram-bot
APScheduler
aiosqlite
python-dotenv
pytz
SQLAlchemy
# Optional: for LLM integration
groq
